requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
